requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.2
django-celery-beat==2.8.0
django-filter==25.1
django-jalali==7.4.0
django-timezone-field==7.1
djangorestframework==3.16.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.3
packaging==25.0
pipdeptree==2.26.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==20.7
redis==6.0.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
